requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp==3.8.1
aiohttp-cors==0.7.0
aioredis==1.3.1
aiorwlock==1.3.0
aiosignal==1.2.0
anyio==3.5.0
arcade==2.6.13
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
blessed==1.19.1
# bzip2==1.0.8
# ca-certificates==2021.10.8
cachetools==5.0.0
# certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorful==0.5.4
cycler==0.11.0
deprecated==1.2.13
dm-tree==0.1.6
fastapi==0.74.1
filelock==3.6.0
fonttools==4.29.1
frozenlist==1.3.0
google-api-core==2.6.0
google-auth==2.6.0
googleapis-common-protos==1.55.0
gpustat==1.0.0b1
gridworld==0.1
grpcio==1.44.0
gym==0.22.0
gym-notices==0.0.4
h11==0.13.0
hiredis==2.0.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.2
jsonschema==4.4.0
kiwisolver==1.3.2
# libffi==3.3
libzlib==1.2.11
lz4==4.0.0
matplotlib==3.5.1
msgpack==1.0.3
multidict==6.0.2
ncurses==6.3
networkx==2.7
numpy==1.22.2
nvidia-ml-py3==7.352.0
opencensus==0.8.0
opencensus-context==0.1.2
openssl==1.1.1l
packaging==21.3
pandas==1.4.1
pillow==9.0.1
pip==22.0.3
prometheus-client==0.13.1
protobuf==3.19.4
psutil==5.9.0
py-spy==0.3.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pygame==2.1.2
pyglet==2.0a2
pymunk==6.2.1
pyparsing==3.0.7
pyrsistent==0.18.1
python==3.9.0
python-dateutil==2.8.2
pytiled-parser==2.0.1
pytz==2021.3
pywavelets==1.2.0
pyyaml==6.0
ray==1.10.0
readline==8.1
redis==4.1.4
requests==2.27.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
setuptools==58.0.4
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sqlite==3.37.0
starlette==0.17.1
tabulate==0.8.9
tensorboardx==2.5
tifffile==2022.2.9
tk==8.6.12
typing-extensions==4.1.1
tzdata==2021e
urllib3==1.26.8
uvicorn==0.17.5
wheel==0.37.1
wrapt==1.13.3
xz==5.2.5
yarl==1.7.2
zipp==3.7.0
zlib==1.2.11
